Like count

+++ src/main/java/uz/smartup/academy/bloggingplatform/service/LikeServiceImpl.java
@@ -19,15 +19,13 @@ public class LikeServiceImpl implements LikeService {
     private final LikeDTOUtil likeDTOUtil;
     private final UserDao userDao;
 
-
     @Autowired
-    public LikeServiceImpl(LikeDAO likeDAO, PostDao postDao, LikeDTOUtil likeDTOUtil, UserService userService, UserDao userDao) {
+    public LikeServiceImpl(LikeDAO likeDAO, PostDao postDao, LikeDTOUtil likeDTOUtil, UserDao userDao) {
         this.likeDAO = likeDAO;
         this.postDao = postDao;
         this.likeDTOUtil = likeDTOUtil;
         this.userDao = userDao;
     }
-
 
     @Override
     public LikeDTO findByUserAndPost(int userId, int postId) {
@@ -36,12 +34,10 @@ public class LikeServiceImpl implements LikeService {
 
     @Override
     public void addLike(int userId, int postId) {
-
         User user = userDao.getUserById(userId);
-
         Post post = postDao.getById(postId);
 
-        if (likeDAO.findByUserAndPost(userId, postId)==null) {
+        if (likeDAO.findByUserAndPost(userId, postId) != null) {
             throw new IllegalStateException("Post already liked");
         }
 
@@ -49,18 +45,22 @@ public class LikeServiceImpl implements LikeService {
         like.setAuthor(user);
         like.setPost(post);
         likeDAO.save(like);
-
     }
 
     @Override
     public void removeLike(int userId, int postId) {
-
         User user = userDao.getUserById(userId);
-
         Post post = postDao.getById(postId);
 
         Like like = likeDAO.findByUserAndPost(userId, postId);
 
-        likeDAO.delete(like);
+        if (like != null) {
+            likeDAO.delete(like);
+        }
+    }
+
+    @Override
+    public long countLikesByPostId(int postId) {
+        return likeDAO.countByPostId(postId);
     }
 }
